1824 Refactored code stages into single file (#313)
* test

* test

* began modifying docstrings

* added parameters

* modified docstrings

* Add docstring for merge functions

* Add docstring for merge functions

* added doc strings to pipeline stages

* Add docstring for calculate_individual_level_population_totals

* Add remaining docstrings

* Add docstring for impute_demographic_columns

* added report doctstring

* Add whitespace to docstring

* added record_level_interface

* Refactor weight calibration stage into sub functions

* Add default parameters for input processing functions

* imports in run.py

* updated to fix complete run

* updated to fix complete run

* fixed imports

* fixed imports

* Add Regenesees install to Actions

* Add R setup to Action

* Separate R based stages

* uncommented barcode cleaining

* Update regression outputs
