+++ .github/workflows/main.yml
@@ -1,6 +1,6 @@
 # This is a basic workflow to help you get started with Actions
 
-name: Testing and coverage
+name: Testing
 
 # Controls when the action will run.
 on:
@@ -29,7 +29,6 @@ jobs:
           python -m pip install --upgrade pip
           pip install wheel==0.29.0
           python setup.py sdist bdist_wheel
-
   test:
     runs-on: ubuntu-latest
     steps:
@@ -61,49 +60,7 @@ jobs:
           python -m pip install --upgrade pip
           pip install pypandoc==1.5
           pip install -e .[ci]
+
       - name: Run pytest
         run: |
           pytest -v
-
-  test-with-cov:
-    runs-on: ubuntu-latest
-    steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE
-      - uses: actions/checkout@v2
-
-      - uses: actions/setup-python@v2
-        with:
-          python-version: 3.6.8
-
-      - uses: actions/setup-java@v1
-        with:
-          java-version: '8'
-
-      - name: Setup spark
-        run: |
-          cd ../../ &&
-          wget -q -O spark.tgz https://archive.apache.org/dist/spark/spark-2.4.0/spark-2.4.0-bin-hadoop2.6.tgz &&
-          tar xzvf spark.tgz &&
-          rm "spark.tgz" &&
-          (cd spark-2.4.0-bin-hadoop2.6 &&
-          export SPARK_HOME=`pwd` &&
-          export PYTHONPATH=$(ZIPS=("$SPARK_HOME"/python/lib/*.zip); IFS=:; echo "${ZIPS[*]}"):$PYTHONPATH
-          )
-
-      - name: Install Python dependencies
-        # Wheel is specific here because of conflicts on this Python version
-        run: |
-          python -m pip install --upgrade pip
-          pip install pypandoc==1.5
-          pip install -e .[ci]
-      - name: Run pytest with coverage
-        run: |
-          pytest --cov=cishouseholds --cov-report=xml --cov-report=term-missing
-          coverage xml
-      - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v1
-        with:
-          name: cishouseholds-pytests-py3.6.8
-          flags: pytests
-          file: ./coverage.xml
-          fail_ci_if_error: true
