+++ cishouseholds/pipeline/generate_outputs.py
@@ -1,178 +1,16 @@
 import subprocess
-from datetime import datetime
-from io import BytesIO
 from pathlib import Path
 from typing import Any
 from typing import List
 from typing import Optional
 from typing import Union
 
-import pandas as pd
-import yaml
 from pyspark.sql import DataFrame
-from pyspark.sql import functions as F
-from pyspark.sql import Window
 
 from cishouseholds.edit import assign_from_map
 from cishouseholds.edit import rename_column_names
 from cishouseholds.edit import update_column_values_from_map
-from cishouseholds.edit import update_from_csv_lookup
 from cishouseholds.extract import list_contents
-from cishouseholds.hdfs_utils import read_header
-from cishouseholds.hdfs_utils import write_string_to_file
-from cishouseholds.pipeline.category_map import category_maps
-from cishouseholds.pipeline.load import check_table_exists
-from cishouseholds.pipeline.load import extract_from_table
-from cishouseholds.pipeline.load import update_table
-from cishouseholds.pipeline.manifest import Manifest
-from cishouseholds.pipeline.pipeline_stages import register_pipeline_stage
-
-
-@register_pipeline_stage("report")
-def report(
-    unique_id_column: str,
-    validation_failure_flag_column: str,
-    duplicate_count_column_name: str,
-    valid_survey_responses_table: str,
-    invalid_survey_responses_table: str,
-    filtered_survey_responses_table: str,
-    output_directory: str,
-):
-    valid_df = extract_from_table(valid_survey_responses_table)
-    invalid_df = extract_from_table(invalid_survey_responses_table)
-    filtered_df = extract_from_table(filtered_survey_responses_table)
-    processed_file_log = extract_from_table("processed_filenames")
-
-    invalid_files_count = 0
-    if check_table_exists("error_file_log"):
-        invalid_files_log = extract_from_table("error_file_log")
-        invalid_files_count = invalid_files_log.count()
-
-    valid_survey_responses_count = valid_df.count()
-    invalid_survey_responses_count = invalid_df.count()
-    filtered_survey_responses_count = filtered_df.count()
-
-    valid_df_errors = valid_df.select(unique_id_column, validation_failure_flag_column)
-    invalid_df_errors = invalid_df.select(unique_id_column, validation_failure_flag_column)
-
-    valid_df_errors = (
-        valid_df_errors.withColumn("Validation check failures", F.explode(validation_failure_flag_column))
-        .groupBy("Validation check failures")
-        .count()
-    )
-    invalid_df_errors = (
-        invalid_df_errors.withColumn("Validation check failures", F.explode(validation_failure_flag_column))
-        .groupBy("Validation check failures")
-        .count()
-    )
-
-    duplicated_df = valid_df.select(unique_id_column, duplicate_count_column_name).filter(
-        F.col(duplicate_count_column_name) > 1
-    )
-    window = Window.partitionBy("run_id")
-    processed_file_log = processed_file_log.withColumn("max", F.max("run_id").over(window))
-    processed_file_log = processed_file_log.filter(F.col("run_id") == F.col("max")).drop("max")
-
-    counts_df = pd.DataFrame(
-        {
-            "dataset": [
-                "invalid input files",
-                "valid survey responses",
-                "invalid survey responses",
-                "filtered survey responses",
-                *list(processed_file_log.select("processed_filename").rdd.flatMap(lambda x: x).collect()),
-            ],
-            "count": [
-                invalid_files_count,
-                valid_survey_responses_count,
-                filtered_survey_responses_count,
-                invalid_survey_responses_count,
-                *list(processed_file_log.select("file_row_count").rdd.flatMap(lambda x: x).collect()),
-            ],
-        }
-    )
-
-    output = BytesIO()
-    with pd.ExcelWriter(output) as writer:
-        counts_df.to_excel(writer, sheet_name="dataset totals", index=False)
-        valid_df_errors.toPandas().to_excel(
-            writer, sheet_name="validation check failures in valid dataset", index=False
-        )
-        invalid_df_errors.toPandas().to_excel(
-            writer, sheet_name="validation check failures in invalid dataset", index=False
-        )
-        duplicated_df.toPandas().to_excel(writer, sheet_name="duplicated record summary", index=False)
-        invalid_files_log.select("file_path", "error").toPandas().to_excel(
-            writer, sheet_name="invalid files summary", index=False
-        )
-
-    write_string_to_file(
-        output.getbuffer(), f"{output_directory}/report_output_{datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}.xlsx"
-    )
-
-
-@register_pipeline_stage("record_level_interface")
-def record_level_interface(
-    survey_responses_table: str,
-    csv_editing_file: str,
-    unique_id_column: str,
-    unique_id_list: List,
-    edited_survey_responses_table: str,
-    filtered_survey_responses_table: str,
-):
-    input_df = extract_from_table(survey_responses_table)
-    edited_df = update_from_csv_lookup(df=input_df, csv_filepath=csv_editing_file, id_column=unique_id_column)
-    update_table(edited_df, edited_survey_responses_table, "overwrite")
-
-    filtered_df = edited_df.filter(F.col(unique_id_column).isin(unique_id_list))
-    update_table(filtered_df, filtered_survey_responses_table, "overwrite")
-
-
-@register_pipeline_stage("tables_to_csv")
-def tables_to_csv(
-    outgoing_directory,
-    tables_to_csv_config_file,
-    category_map,
-    dry_run=False,
-):
-    """
-    Parameters
-    ----------
-    outgoing_directory
-    tables_to_csv_config_file
-    category_map
-    dry_run
-    use_table_to_csv_config
-
-    Writes data from an existing HIVE table to csv output, including mapping of column names and values.
-    Takes a yaml file in which Hive table name and csv table name are defined as well as columns to be
-    included in the csv file by a select statement.
-    Optionally also point to an update map to be used for the variable name mapping of these outputs.
-    """
-    output_datetime = datetime.today()
-    output_datetime_str = output_datetime.strftime("%Y%m%d_%H%M%S")
-
-    file_directory = Path(outgoing_directory) / output_datetime_str
-    manifest = Manifest(outgoing_directory, pipeline_run_datetime=output_datetime, dry_run=dry_run)
-    category_map_dictionary = category_maps[category_map]
-
-    with open(tables_to_csv_config_file) as f:
-        config_file = yaml.load(f, Loader=yaml.FullLoader)
-
-    for table in config_file["create_tables"]:
-        df = extract_from_table(table["table_name"]).select(*[element for element in table["column_name_map"].keys()])
-        df = map_output_values_and_column_names(df, table["column_name_map"], category_map_dictionary)
-        file_path = file_directory / f"{table['output_file_name']}_{output_datetime_str}"
-        write_csv_rename(df, file_path)
-        file_path = file_path.with_suffix(".csv")
-        header_string = read_header(file_path)
-
-        manifest.add_file(
-            relative_file_path=file_path.relative_to(outgoing_directory).as_posix(),
-            column_header=header_string,
-            validate_col_name_length=False,
-        )
-    manifest.write_manifest()
 
 
 def map_output_values_and_column_names(df: DataFrame, column_name_map: dict, value_map_by_column: dict):
