+++ cishouseholds/pipeline/input_file_processing.py
@@ -1,8 +1,6 @@
-from cishouseholds.extract import get_files_to_be_processed
 from cishouseholds.pipeline.blood_delta_ETL import add_historical_fields
 from cishouseholds.pipeline.blood_delta_ETL import transform_blood_delta
 from cishouseholds.pipeline.cast_columns_from_string_map import survey_response_cast_to_double
-from cishouseholds.pipeline.ETL_scripts import extract_validate_transform_input_data
 from cishouseholds.pipeline.historical_blood_ETL import add_fields
 from cishouseholds.pipeline.input_variable_names import blood_variable_name_map
 from cishouseholds.pipeline.input_variable_names import historical_blood_variable_name_map
@@ -11,8 +9,7 @@ from cishouseholds.pipeline.input_variable_names import survey_responses_v1_vari
 from cishouseholds.pipeline.input_variable_names import survey_responses_v2_variable_name_map
 from cishouseholds.pipeline.input_variable_names import swab_variable_name_map
 from cishouseholds.pipeline.input_variable_names import unassayed_bloods_variable_name_map
-from cishouseholds.pipeline.load import update_table_and_log_source_files
-from cishouseholds.pipeline.pipeline_stages import register_pipeline_stage
+from cishouseholds.pipeline.pipeline_stages import generate_input_processing_function
 from cishouseholds.pipeline.survey_responses_version_0_ETL import transform_survey_responses_version_0_delta
 from cishouseholds.pipeline.survey_responses_version_1_ETL import transform_survey_responses_version_1_delta
 from cishouseholds.pipeline.survey_responses_version_2_ETL import derive_additional_v1_2_columns
@@ -34,7 +31,6 @@ from cishouseholds.pipeline.validation_schema import survey_responses_v2_validat
 from cishouseholds.pipeline.validation_schema import swab_validation_schema
 from cishouseholds.pipeline.validation_schema import swab_validation_schema_testKit
 from cishouseholds.pipeline.validation_schema import unassayed_blood_validation_schema
-from cishouseholds.validate import validate_files
 
 
 blood_delta_parameters = {
@@ -137,66 +133,6 @@ historical_blood_parameters = {
     "output_table_name": "transformed_blood_test_data",
     "source_file_column": "blood_test_source_file",
 }
-
-
-def generate_input_processing_function(
-    stage_name,
-    validation_schema,
-    column_name_map,
-    datetime_column_map,
-    transformation_functions,
-    output_table_name,
-    source_file_column,
-    write_mode="overwrite",
-    sep=",",
-    cast_to_double_list=[],
-    include_hadoop_read_write=True,
-):
-    """
-    Generate an input file processing stage function and register it.
-
-    Returns dataframe for use in testing.
-
-    Parameters
-    ----------
-    include_hadoop_read_write
-        set to False for use in testing on non-hadoop environments
-
-    Notes
-    -----
-    See underlying functions for other parameter documentation.
-    """
-
-    @register_pipeline_stage(stage_name)
-    def _inner_function(**kwargs):
-        file_path_list = [kwargs["resource_path"]]
-        if include_hadoop_read_write:
-            file_path_list = get_files_to_be_processed(**kwargs)
-        if not file_path_list:
-            print(f"        - No files selected in {kwargs['resource_path']}")  # functional
-            return
-
-        valid_file_paths = validate_files(file_path_list, validation_schema, sep=sep)
-        if not valid_file_paths:
-            print(f"        - No valid files found in: {kwargs['resource_path']}.")  # functional
-            return
-
-        df = extract_validate_transform_input_data(
-            resource_path=file_path_list,
-            variable_name_map=column_name_map,
-            datetime_map=datetime_column_map,
-            validation_schema=validation_schema,
-            transformation_functions=transformation_functions,
-            sep=sep,
-            cast_to_double_columns_list=cast_to_double_list,
-        )
-        if include_hadoop_read_write:
-            update_table_and_log_source_files(df, output_table_name, source_file_column, write_mode)
-        return df
-
-    _inner_function.__name__ = stage_name
-    return _inner_function
-
 
 for parameters in [
     blood_delta_parameters,
