+++ cishouseholds/pipeline/run.py
@@ -1,21 +1,13 @@
 import traceback
 from datetime import datetime
 
-import cishouseholds.pipeline.generate_outputs  # noqa: F401
-import cishouseholds.pipeline.historical_blood_ETL  # noqa: F401
 import cishouseholds.pipeline.input_file_processing  # noqa: F401
-import cishouseholds.pipeline.merge_antibody_swab_ETL  # noqa: F401
-import cishouseholds.pipeline.post_merge_processing  # noqa: F401
-import cishouseholds.pipeline.survey_responses_version_0_ETL  # noqa: F401
-import cishouseholds.pipeline.survey_responses_version_1_ETL  # noqa: F401
-import cishouseholds.pipeline.survey_responses_version_2_ETL  # noqa: F401
-import cishouseholds.pipeline.weight_calibration  # noqa: F401
+import cishouseholds.pipeline.pipeline_stages  # noqa: F401
+import cishouseholds.pipeline.R_pipeline_stages  # noqa: F401
 from cishouseholds.pipeline.config import get_config
 from cishouseholds.pipeline.load import add_run_log_entry
 from cishouseholds.pipeline.load import add_run_status
-from cishouseholds.pipeline.load import delete_tables  # noqa: F401
 from cishouseholds.pipeline.pipeline_stages import pipeline_stages
-from dummy_data_generation.generate_data import generate_dummy_data  # noqa: F401
 
 
 def run_from_config():
