+++ tests/pipeline/test_blood_deltas/processed_blood_data_regression.yml
@@ -11,7 +11,7 @@
   antibody_test_well_id: O25
   blood_sample_arrayed_date: '2018-11-24 00:00:00.000000'
   blood_sample_barcode: ONS00227680
-  blood_sample_barcode_edited_flag: 1
+  blood_sample_barcode_edited_in_bloods_dataset_flag: 1
   blood_sample_collected_datetime: '2018-09-07 22:41:44.000000'
   blood_sample_received_date: '2018-01-03 00:00:00.000000'
   blood_sample_type: Venous
@@ -33,7 +33,7 @@
   antibody_test_well_id: Z35
   blood_sample_arrayed_date: '2018-03-23 00:00:00.000000'
   blood_sample_barcode: ONS00863193
-  blood_sample_barcode_edited_flag: 1
+  blood_sample_barcode_edited_in_bloods_dataset_flag: 1
   blood_sample_collected_datetime: '2020-10-09 01:46:29.000000'
   blood_sample_received_date: '2019-06-04 00:00:00.000000'
   blood_sample_type: Capillary
@@ -55,7 +55,7 @@
   antibody_test_well_id: N00
   blood_sample_arrayed_date: '2020-02-07 00:00:00.000000'
   blood_sample_barcode: ONS03639422
-  blood_sample_barcode_edited_flag: 1
+  blood_sample_barcode_edited_in_bloods_dataset_flag: 1
   blood_sample_collected_datetime: '2021-03-09 04:15:47.000000'
   blood_sample_received_date: '2019-11-16 00:00:00.000000'
   blood_sample_type: Venous
@@ -77,7 +77,7 @@
   antibody_test_well_id: G29
   blood_sample_arrayed_date: '2018-10-21 00:00:00.000000'
   blood_sample_barcode: ONS06171396
-  blood_sample_barcode_edited_flag: 1
+  blood_sample_barcode_edited_in_bloods_dataset_flag: 1
   blood_sample_collected_datetime: '2022-04-22 10:53:49.000000'
   blood_sample_received_date: '2019-08-13 00:00:00.000000'
   blood_sample_type: Venous
@@ -99,7 +99,7 @@
   antibody_test_well_id: F29
   blood_sample_arrayed_date: '2022-06-19 00:00:00.000000'
   blood_sample_barcode: ONS06857022
-  blood_sample_barcode_edited_flag: 1
+  blood_sample_barcode_edited_in_bloods_dataset_flag: 1
   blood_sample_collected_datetime: '2020-08-21 11:10:23.000000'
   blood_sample_received_date: '2019-05-03 00:00:00.000000'
   blood_sample_type: Venous
